SONAR-9991 Set BranchPersister as Inteface
This will allow privileged plugin as GOV to do nothing on branch
